maldade agora é int

+++ include/Entidades/Personagens/Inimigo.h
@@ -13,7 +13,7 @@ namespace Alaska
             class Inimigo : public Personagem
             {
                 protected:
-                    float nivel_maldade;
+                    int nivel_maldade;
                     Jogador* pJogador;
                     float velocidade;
                 protected:
@@ -27,7 +27,7 @@ namespace Alaska
                     virtual void salvar() = 0;
                     virtual void mover() = 0;
 
-                    float calcularMaldade();
+                    int calcularMaldade();
                     int getMaldade();
                     sf::Vector2f normalizarVetor(sf::Vector2f v);
 					void salvarDataBuffer();
